Move the toolchain configuration into the "apache-release" profile
For building releases in Apache this configuration is quite relevant: we don't
want to release something that claims 1.6 compatibility but requires a higher
version.

For our fork though this is not relevant: we're fine with Java 8, and having
to configure build machines with weird toolchains is not really needed (and
in fact just complicates matters).

NB: samples-spring-client doesn't build with Java 8+ due to a bug in the webstart
plugin, which is already fixed in a recent snapshot.